Add GPU-required tests.
This commit implements the --gpu flag for running tests. When set
like with `uv run pytest --gpu` tests marked with @mark_gpu will
be run in addition to other default tests.

This setup was done instead of using marks because it sets the
default behavior to not run gpu-required tests (making the tests
work by default on most systems). Running GPU-tests is opt-in
and can be setup to be executed by adding the flag when desired.